invoice: move exchange rate API from fui to io module
Also switch from using `requests` lib to `urllib.request` (standard python library) and improve error popups.
