+++ electroncash_gui/qt/invoice_dialog.py
@@ -25,17 +25,19 @@
 from __future__ import annotations
 
 import json
-from dataclasses import dataclass
 from decimal import Decimal
-from typing import List, Optional, Sequence, Union
+from http.client import InvalidURL
+from typing import List, Optional, Union
+from urllib.error import URLError
 
-import requests
 from PyQt5 import QtCore, QtWidgets
 
 from electroncash.address import Address, AddressError
 from electroncash.i18n import _
 from electroncash.invoice import (
+    APIS,
     APIExchangeRate,
+    ExchangeRateAPI,
     FixedExchangeRate,
     Invoice,
     InvoiceDataError,
@@ -163,7 +165,7 @@ class InvoiceDialog(QtWidgets.QDialog):
                     _("Unable to decode JSON data for invoice")
                     + f" {filename}.\n\n"
                     + _("The following error was raised:")
-                    + str(e),
+                    + f"{type(e).__name__}: {e}",
                 )
                 return
         try:
@@ -174,8 +176,8 @@ class InvoiceDialog(QtWidgets.QDialog):
                 _("Failed to load invoice"),
                 _("Unable to decode JSON data for invoice")
                 + f" {filename}.\n\n"
-                + _("The following error was raised:")
-                + str(e),
+                + _("The following error was raised:\n\n")
+                + f"{type(e).__name__}: {e}",
             )
             return
 
@@ -304,59 +306,6 @@ class ExchangeRateWidget(QtWidgets.QWidget):
         self.rate_edit.setValue(1.0)
 
 
-@dataclass
-class ExchangeRateAPI:
-    url: str
-    keys: Sequence[str]
-
-    def get_url(self, currency: str) -> str:
-        """Get request url with occurrences of ${cur} and %CUR% replaced with
-        respectively lower case or upper case currency symbol.
-        """
-        url = self.url.replace("%cur%", currency.lower())
-        return url.replace("%CUR%", currency.upper())
-
-    def get_keys(self, currency: str) -> List[str]:
-        """Get keys with occurrences of %cur% and %CUR% replaced with
-        respectively lower case or upper case currency symbol.
-        """
-        return [
-            k.replace("%cur%", currency.lower()).replace("%CUR%", currency.upper())
-            for k in self.keys
-        ]
-
-    def get_exchange_rate(self, currency: str) -> float:
-        url = self.get_url(currency)
-        keys = self.get_keys(currency)
-
-        json_data = requests.get(url).json()
-
-        next_node = json_data
-        for k in keys:
-            next_node = next_node[k]
-        return float(next_node)
-
-
-APIS: List[ExchangeRateAPI] = [
-    ExchangeRateAPI(
-        "https://api.coingecko.com/api/v3/simple/price?ids=ecash&vs_currencies=%cur%",
-        ["ecash", "%cur%"],
-    ),
-    ExchangeRateAPI(
-        "https://api.coingecko.com/api/v3/coins/ecash?localization=False&sparkline=false",
-        ["market_data", "current_price", "%cur%"],
-    ),
-    ExchangeRateAPI(
-        "https://api.binance.com/api/v3/avgPrice?symbol=XECUSDT",
-        ["price"],
-    ),
-    ExchangeRateAPI(
-        "https://api.binance.com/api/v3/avgPrice?symbol=XECBUSD",
-        ["price"],
-    ),
-]
-
-
 class ExchangeRateAPIWidget(QtWidgets.QWidget):
     def __init__(self, parent: Optional[QtWidgets.QWidget] = None):
         super().__init__(parent)
@@ -421,12 +370,20 @@ class ExchangeRateAPIWidget(QtWidgets.QWidget):
         api = ExchangeRateAPI(self.get_url(), self.get_keys())
         try:
             rate = api.get_exchange_rate(self._currency)
-        except (KeyError, requests.exceptions.RequestException) as e:
+        except (
+            # Wrong currency, currency key not available for this API
+            KeyError,
+            # urllib raises ValueError on unrecognized url types
+            ValueError,
+            URLError,
+            InvalidURL,
+        ) as e:
             QtWidgets.QMessageBox.critical(
                 self,
                 "Error fetching exchange rate",
                 f"Unable to fetch the XEC/{self._currency} exchange rate using the "
-                f"specified API parameters.\n\nThe error message was:\n\n{e}",
+                f"specified API parameters.\n\nThe error message was:\n\n"
+                f"{type(e).__name__}: {e}",
             )
             return
